requirements.txt
Django==4.2.5
django-environ==0.11.2
psycopg==3.1.12
django-basicauth==0.5.3
